pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mark"
version = "1.0.0"
description = "Versatile agents for multitasking with livekit and python"
requires-python = ">=3.10, <3.15"

dependencies = [
    "httpx>=0.28.1",
    "livekit-agents[silero,turn-detector]~=1.4",
    "livekit-plugins-ai-coustics",
    "livekit-plugins-noise-cancellation~=0.2",
    "playwright>=1.58.0",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]


[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
